Version 0.9.0 - Rewritten webpack code to use esm - Refactored webpack code to be more modular - Prettier v3 - Pixi v7.3 - Patch Pixi to not export Url - Added Ping endpoint in the API - Babel config from json to code - Changed jsdoc import to be esm imports instead of jsdoc typedefs - Updated nuget packages - Updated npm packages
