+++ src/client/.yarnrc
@@ -1,0 +1,1 @@
+workspaces-experimental false
