+++ src/client/client.esproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.VisualStudio.JavaScript.Sdk/0.5.76-alpha">
+<Project Sdk="Microsoft.VisualStudio.JavaScript.Sdk/0.5.94-alpha">
   <PropertyGroup Label="Globals">
     <ProjectGuid>877dcf29-c85a-4571-a659-118df32af090</ProjectGuid>
   </PropertyGroup>
