+++ src/client/package.json
@@ -6,46 +6,51 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "webpack --config webpack.config.development.cjs",
-    "production": "webpack --config webpack.config.production.cjs",
-    "start": "webpack serve --config webpack.config.development.cjs --open",
+    "build": "webpack --config webpack.development.mjs",
+    "production": "webpack --config webpack.production.mjs",
+    "start": "webpack serve --config webpack.development.mjs --open",
+    "postinstall": "patch-package",
     "prettier-watch": "onchange -i \"src/**/*\" \"*\" -- prettier --write --ignore-unknown {{changed}}",
     "browserslist": "browserslist --env=all"
   },
   "devDependencies": {
-    "@babel/core": "7.22.5",
-    "@babel/eslint-parser": "7.22.5",
-    "@babel/preset-env": "7.22.5",
-    "babel-loader": "9.1.2",
+    "@babel/core": "7.22.9",
+    "@babel/eslint-parser": "7.22.9",
+    "@babel/plugin-syntax-import-assertions": "7.22.5",
+    "@babel/preset-env": "7.22.9",
+    "babel-loader": "9.1.3",
     "bytes": "3.1.2",
     "copy-webpack-plugin": "11.0.0",
     "csp-html-webpack-plugin": "5.1.0",
     "css-loader": "6.8.1",
     "css-minimizer-webpack-plugin": "5.0.1",
     "enhanced-resolve": "5.15.0",
-    "eslint": "8.43.0",
+    "eslint": "8.44.0",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-compat": "4.1.4",
     "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-jsdoc": "46.4.2",
+    "eslint-plugin-jsdoc": "46.4.3",
     "eslint-plugin-no-unsanitized": "4.0.2",
-    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-prettier": "5.0.0",
     "eslint-webpack-plugin": "4.0.1",
-    "glob": "10.3.1",
+    "glob": "10.3.3",
     "html-inline-css-webpack-plugin": "1.11.1",
     "html-webpack-plugin": "5.5.3",
     "ignore-loader": "0.1.2",
     "immutability-helper": "3.1.1",
     "mini-css-extract-plugin": "2.7.6",
-    "minimatch": "9.0.2",
+    "minimatch": "9.0.3",
     "onchange": "7.1.0",
-    "postcss": "8.4.24",
+    "patch-package": "7.0.2",
+    "postcss": "8.4.26",
+    "postcss-load-config": "4.0.1",
     "postcss-loader": "7.3.3",
     "postcss-merge-rules": "6.0.1",
-    "postcss-preset-env": "8.5.1",
-    "prettier": "2.8.8",
-    "stylelint": "15.9.0",
-    "stylelint-config-standard": "33.0.0",
+    "postcss-preset-env": "9.0.0",
+    "postinstall-postinstall": "2.1.0",
+    "prettier": "3.0.0",
+    "stylelint": "15.10.1",
+    "stylelint-config-standard": "34.0.0",
     "stylelint-webpack-plugin": "4.1.1",
     "tslib": "2.6.0",
     "typescript": "5.1.6",
@@ -56,23 +61,15 @@
   },
   "dependencies": {
     "@microsoft/applicationinsights-web": "3.0.2",
-    "@pixi/app": "7.2.4",
-    "@pixi/core": "7.2.4",
-    "@pixi/display": "7.2.4",
-    "@pixi/graphics": "7.2.4",
-    "@pixi/sprite": "7.2.4",
-    "@pixi/unsafe-eval": "7.2.4",
-    "@pixi/utils": "7.2.4",
+    "@pixi/app": "7.3.0-rc",
+    "@pixi/core": "7.3.0-rc",
+    "@pixi/display": "7.3.0-rc",
+    "@pixi/graphics": "7.3.0-rc",
+    "@pixi/sprite": "7.3.0-rc",
+    "@pixi/unsafe-eval": "7.3.0-rc",
+    "@pixi/utils": "7.3.0-rc",
     "core-js": "3.31.1",
     "navigator.sendbeacon": "0.0.20",
     "unfetch": "5.0.0"
-  },
-  "resolutions": {
-    "**/@pixi/utils/url": "0.11.0"
-  },
-  "overrides": {
-    "@pixi/utils": {
-      "url": "0.11.0"
-    }
   }
 }
