+++ src/client/settings.mjs
@@ -1,0 +1,8 @@
+export default {
+  development: {
+    INSTRUMENTATION_KEY: '1ef798fd-96b6-4e8b-a8c0-ea41a23d05b8'
+  },
+  production: {
+    INSTRUMENTATION_KEY: 'efd49a11-d1f6-4907-a112-40e07bcd7f56'
+  }
+};
